refactor: extract config provider ownership
